Move AppWorker and add WorkConstants
