+++ app/src/main/java/org/cssnr/noaaweather/work/WorkConstants.kt
@@ -1,0 +1,11 @@
+package org.cssnr.noaaweather.work
+
+import androidx.work.Constraints
+import androidx.work.NetworkType
+
+val APP_WORKER_CONSTRAINTS: Constraints = Constraints.Builder()
+    .setRequiresBatteryNotLow(true)
+    .setRequiresCharging(false)
+    .setRequiresDeviceIdle(false)
+    .setRequiredNetworkType(NetworkType.NOT_REQUIRED)
+    .build()
